django_app/requirements.txt
Django==4.2.7
PyMySQL==1.1.1
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
honeybadger @ git+https://github.com/honeybadger-io/honeybadger-python.git@insights-instrumentation#egg=honeybadger
